+++ shenyu-bootstrap/src/main/resources/application.yml
@@ -148,7 +148,7 @@ shenyu:
 #      workerCount: 8
 #      daemon: true
   register:
-    enabled: true
+    enabled: false
     registerType: zookeeper #etcd #consul
     serverLists: localhost:2181 #http://localhost:2379 #localhost:8848
     props:
