+++ shenyu-examples/shenyu-examples-sdk/shenyu-examples-sdk-dubbo/shenyu-examples-sdk-apache-dubbo-consumer/pom.xml
@@ -25,7 +25,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>shenyu-examples-sdk-apache-dubbo</artifactId>
+    <artifactId>shenyu-examples-sdk-apache-dubbo-consumer</artifactId>
 
     <dependencies>
         <dependency>
@@ -35,6 +35,7 @@
         <dependency>
             <groupId>org.apache.shenyu</groupId>
             <artifactId>shenyu-examples-dubbo-api</artifactId>
+            <version>${project.version}</version>
         </dependency>
 
         <dependency>
@@ -70,13 +71,13 @@
     </dependencies>
 
     <build>
-        <finalName>shenyu-examples-sdk-http</finalName>
+        <finalName>shenyu-examples-sdk-apache-dubbo-consumer</finalName>
         <plugins>
             <plugin>
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-maven-plugin</artifactId>
                 <configuration>
-                    <mainClass>org.apache.shenyu.examples.sdk.apache.dubbo.ShenyuSdkHttpExampleApplication</mainClass>
+                    <mainClass>org.apache.shenyu.examples.sdk.apache.dubbo.consumer.ShenyuSdkApacheDubboExampleApplication</mainClass>
                     <executable>true</executable>
                 </configuration>
             </plugin>
@@ -86,8 +87,8 @@
         <profile>
             <id>example</id>
             <properties>
-                <docker.buildArg.APP_NAME>shenyu-examples-sdk-http</docker.buildArg.APP_NAME>
-                <docker.image.tag.repo>shenyu-examples-sdk-http</docker.image.tag.repo>
+                <docker.buildArg.APP_NAME>shenyu-examples-sdk-apache-dubbo-consumer</docker.buildArg.APP_NAME>
+                <docker.image.tag.repo>shenyu-examples-sdk-apache-dubbo-consumer</docker.image.tag.repo>
                 <docker.image.tag.tagName>latest</docker.image.tag.tagName>
             </properties>
             <activation>
@@ -102,7 +103,7 @@
                         <configuration>
                             <images>
                                 <image>
-                                    <name>shenyu-examples-sdk-http</name>
+                                    <name>shenyu-examples-sdk-apache-dubbo-consumer</name>
                                     <build>
                                         <contextDir>${project.basedir}</contextDir>
                                     </build>
